# public/swagger/documentation.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "API de Gestion des Étudiants",
      "version": "1.0.0",
      "description": "API pour gérer les étudiants, leurs notes et l'authentification des utilisateurs dans une école."
    },
    "servers": [
      {
        "url": "http://127.0.0.1:8000/api",
        "description": "Serveur de développement"
      }
    ],
    "paths": {
      "/etudiants": {
        "get": {
          "summary": "Liste des étudiants",
          "tags": ["Étudiants"],
          "security": [
            {
              "BearerAuth": []
            }
          ],
          "responses": {
            "200": {
              "description": "Liste des étudiants",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Etudiant"
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "summary": "Créer un nouvel étudiant",
          "tags": ["Étudiants"],
          "security": [
            {
              "BearerAuth": []
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EtudiantCreate"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Étudiant créé avec succès",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Etudiant"
                  }
                }
              }
            }
          }
        }
      },
      "/etudiants/{id}": {
        "get": {
          "summary": "Récupérer un étudiant",
          "tags": ["Étudiants"],
          "security": [
            {
              "BearerAuth": []
            }
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "description": "ID de l'étudiant",
              "schema": {
                "type": "integer"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Étudiant récupéré avec succès",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Etudiant"
                  }
                }
              }
            }
          }
        },
        "post": {
          "summary": "Modifier un étudiant",
          "tags": ["Étudiants"],
          "security": [
            {
              "BearerAuth": []
            }
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "description": "ID de l'étudiant",
              "schema": {
                "type": "integer"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EtudiantUpdate"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Étudiant modifié avec succès",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Etudiant"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "summary": "Supprimer un étudiant",
          "tags": ["Étudiants"],
          "security": [
            {
              "BearerAuth": []
            }
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "description": "ID de l'étudiant",
              "schema": {
                "type": "integer"
              }
            }
          ],
          "responses": {
            "204": {
              "description": "Étudiant supprimé avec succès"
            }
          }
        }
      },
      "/login": {
        "post": {
          "summary": "Connexion utilisateur",
          "tags": ["Authentification"],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "type": "string",
                      "format": "email",
                      "example": "admin@example.com"
                    },
                    "password": {
                      "type": "string",
                      "example": "password123"
                    }
                  },
                  "required": ["email", "password"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Connexion réussie",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "access_token": {
                        "type": "string",
                        "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
                      },
                      "token_type": {
                        "type": "string",
                        "example": "bearer"
                      },
                      "user": {
                        "$ref": "#/components/schemas/User"
                      },
                      "expires_in": {
                        "type": "string",
                        "example": "3600 seconds"
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Informations de connexion incorrectes",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string",
                        "example": "Informations de connexion incorrectes"
                      }
                    }
                  }
                }
              }
            },
            "422": {
              "description": "Validation des données échouée",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "errors": {
                        "type": "object",
                        "additionalProperties": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/logout": {
        "post": {
          "summary": "Déconnexion utilisateur",
          "tags": ["Authentification"],
          "responses": {
            "200": {
              "description": "Déconnexion réussie",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string",
                        "example": "Déconnexion réussie"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/refresh": {
        "post": {
          "summary": "Rafraîchissement du token",
          "tags": ["Authentification"],
          "responses": {
            "200": {
              "description": "Token rafraîchi avec succès",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "access_token": {
                        "type": "string",
                        "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
                      },
                      "token_type": {
                        "type": "string",
                        "example": "bearer"
                      },
                      "user": {
                        "$ref": "#/components/schemas/User"
                      },
                      "expires_in": {
                        "type": "string",
                        "example": "3600 seconds"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "Etudiant": {
          "type": "object",
          "properties": {
            "prenom": {
              "type": "string",
              "example": "Jean"
            },
            "nom": {
              "type": "string",
              "example": "Dupont"
            },
            "adresse": {
              "type": "string",
              "example": "10 rue des Lilas"
            },
            "telephone": {
              "type": "integer",
              "example": 123456789
            },
            "date_de_naissance": {
              "type": "string",
              "format": "date",
              "example": "2000-01-01"
            },
            "matricule": {
              "type": "string",
              "example": "2020001"
            },
            "email": {
              "type": "string",
              "format": "email",
              "example": "jean.dupont@example.com"
            }
          },
          "required": ["prenom", "nom", "adresse", "telephone", "date_de_naissance", "matricule", "email"]
        },
        "EtudiantCreate": {
          "type": "object",
          "properties": {
            "prenom": {
              "type": "string",
              "example": "Jean"
            },
            "nom": {
              "type": "string",
              "example": "Dupont"
            },
            "adresse": {
              "type": "string",
              "example": "10 rue des Lilas"
            },
            "telephone": {
              "type": "integer",
              "example": 123456789
            },
            "date_de_naissance": {
              "type": "string",
              "format": "date",
              "example": "2000-01-01"
            },
            "matricule": {
              "type": "string",
              "example": "2020001"
            },
            "email": {
              "type": "string",
              "format": "email",
              "example": "jean.dupont@example.com"
            }
          },
          "required": ["prenom", "nom", "adresse", "telephone", "date_de_naissance", "matricule", "email"]
        },
        "EtudiantUpdate": {
          "type": "object",
          "properties": {
            "prenom": {
              "type": "string",
              "example": "Jean"
            },
            "nom": {
              "type": "string",
              "example": "Dupont"
            },
            "adresse": {
              "type": "string",
              "example": "10 rue des Lilas"
            },
            "telephone": {
              "type": "integer",
              "example": 123456789
            },
            "date_de_naissance": {
              "type": "string",
              "format": "date",
              "example": "2000-01-01"
            },
            "matricule": {
              "type": "string",
              "example": "2020001"
            },
            "email": {
              "type": "string",
              "format": "email",
              "example": "jean.dupont@example.com"
            }
          }
        },
        "User": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "example": 1
            },
            "name": {
              "type": "string",
              "example": "John Doe"
            },
            "email": {
              "type": "string",
              "format": "email",
              "example": "john.doe@example.com"
            }
          }
        }
      },
      "securitySchemes": {
        "BearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      }
    }
  }